requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
torch==2.0.1
faiss-cpu==1.7.4
Pillow==9.5.0
tqdm==4.65.0
ftfy==6.1.1
regex==2023.5.5
clip @ git+https://github.com/openai/CLIP.git